requirements.txt
PyMuPDF>=1.24.0
openpyxl>=3.1.0
streamlit>=1.30.0
openai>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
